backend/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Database
sqlalchemy==2.0.36
aiosqlite==0.20.0
asyncpg==0.29.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Redis
redis==5.0.8

# Validation / settings
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic[email]==2.9.2
email-validator==2.2.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
anyio==4.6.0
